REF: break out utils from cases
